# sekre-backend/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Sekre Backend API
  description: |
    Multi-tenant SaaS platform for organization management with task tracking, 
    event scheduling, and financial management.

    ## Features
    - **Organization Management**: Multi-tenant architecture with subscription plans
    - **User Management**: Role-based access control (OWNER, ADMIN, MEMBER)
    - **Division Structure**: Hierarchical team organization
    - **Task Management**: Assign and track tasks with priorities and statuses
    - **Event Scheduling**: Plan and manage organizational events
    - **Finance Tracking**: Multi-currency transaction management with approval workflow
    - **Audit Logging**: Track all critical operations

    ## Authentication Flow

    ### Registration
    1. POST `/auth/register` with organization and user details
    2. System creates organization with FREE plan
    3. First user is assigned OWNER role
    4. Returns access token (1 hour) and refresh token (7 days)

    ### Login
    1. POST `/auth/login` with email and password
    2. Returns access token (1 hour) and refresh token (7 days)
    3. Use access token in `Authorization: Bearer <token>` header

    ### Token Refresh
    1. POST `/auth/refresh` with refresh token
    2. Returns new access token
    3. Refresh token remains valid until expiry

    ### JWT Token Structure
    ```json
    {
      "header": {
        "alg": "HS256",
        "typ": "JWT"
      },
      "payload": {
...
        "role": "OWNER",
...

...

...
servers:
...
security:
  - BearerAuth: []
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new organization
      description: |
        Creates a new organization with the first user as Owner.
        
        **Registration Flow:**
        1. Submit organization and user details
        2. System creates organization with FREE plan
        3. First user is assigned OWNER role
        4. Returns access and refresh tokens
        
        **Password Requirements:**
        - Minimum 8 characters
        - At least one uppercase letter
        - At least one lowercase letter
        - At least one number
        - At least one special character
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            examples:
              basic:
                summary: Basic registration
                value:
                  organization_name: "Acme Corporation"
                  email: "admin@acme.com"
                  password: "SecurePass123!"
                  full_name: "John Doe"
      responses:
        '201':
          description: Organization registered successfully
          headers:
            X-Request-ID:
              schema:
                type: string
                format: uuid
              description: Unique request identifier for tracing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSuccessResponse'
              examples:
                success:
                  summary: Successful registration
                  value:
                    access_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    refresh_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    token_type: "Bearer"
                    expires_in: 3600
                    user:
                      id: "550e8400-e29b-41d4-a716-446655440000"
                      email: "admin@acme.com"
                      full_name: "John Doe"
                      role: "OWNER"
                      organization_id: "660e8400-e29b-41d4-a716-446655440000"
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalid_email:
                  summary: Invalid email format
                  value:
                    error:
                      code: "INVALID_INPUT"
                      message: "Invalid email format"
                      details:
                        - field: "email"
                          message: "Must be a valid email address"
                weak_password:
                  summary: Weak password
                  value:
                    error:
                      code: "INVALID_INPUT"
                      message: "Password does not meet requirements"
                      details:
                        - field: "password"
                          message: "Password must be at least 8 characters with uppercase, lowercase, number, and special character"
                duplicate_email:
                  summary: Email already exists
                  value:
                    error:
                      code: "DUPLICATE_ENTRY"
                      message: "Email already registered"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                server_error:
                  summary: Server error
                  value:
                    error:
                      code: "INTERNAL_ERROR"
                      message: "An unexpected error occurred"
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Subdomain already taken
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          $ref: '#/components/responses/RateLimited'
  /auth/login:
    post:
      tags: [Auth]
      summary: Login
      description: |
        Authenticates with email and password, returns JWT tokens.
        
        **Login Flow:**
        1. Submit email and password
        2. System validates credentials
        3. Returns access token (1 hour expiry) and refresh token (7 days expiry)
        4. Use access token in Authorization header for subsequent requests
        
        **Token Structure:**
        - Access Token: Short-lived (1 hour), used for API requests
        - Refresh Token: Long-lived (7 days), used to obtain new access tokens
        
        **Rate Limiting:**
        - 5 requests per minute per IP
        - Returns 429 with Retry-After header when exceeded
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              basic:
                summary: Basic login
                value:
                  email: "admin@acme.com"
                  password: "SecurePass123!"
      responses:
        '200':
          description: Login successful
          headers:
            X-Request-ID:
              schema:
                type: string
                format: uuid
              description: Unique request identifier for tracing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSuccessResponse'
              examples:
                success:
                  summary: Successful login
                  value:
                    access_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    refresh_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    token_type: "Bearer"
                    expires_in: 3600
                    user:
                      id: "550e8400-e29b-41d4-a716-446655440000"
                      email: "admin@acme.com"
                      full_name: "John Doe"
                      role: "OWNER"
                      organization_id: "660e8400-e29b-41d4-a716-446655440000"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalid_credentials:
                  summary: Invalid email or password
                  value:
                    error:
                      code: "UNAUTHORIZED"
                      message: "Invalid email or password"
        '429':
          $ref: '#/components/responses/RateLimited'
  /auth/me:
    get:
      tags: [Auth]
      summary: Get current user
      description: Returns the authenticated user with their organization.
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserWithOrgResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access and refresh token pair
      description: |
        Rotates refresh token and returns a new token pair.

        **Behavior:**
        - Validates refresh token signature and token type
        - Validates refresh token session in database
        - Revokes old refresh token session
        - Creates a new refresh token session
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
            examples:
              basic:
                value:
                  refresh_token: "eyJhbGciOi..."
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPairResponse'
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/logout:
    post:
      tags: [Auth]
      summary: Logout current user
      description: |
        Revokes all active refresh sessions for the authenticated user.
        Access token must be provided via Bearer auth.
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Invalid request payload
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            validation_error:
              summary: Validation error
              value:
                error:
                  code: "INVALID_INPUT"
                  message: "Validation failed"
                  details:
                    - field: "email"
                      message: "Must be a valid email address"
                    - field: "password"
                      message: "Must be at least 8 characters"
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            missing_token:
              summary: Missing authentication token
              value:
                error:
                  code: "UNAUTHORIZED"
                  message: "Authentication required"
            invalid_token:
              summary: Invalid or expired token
              value:
                error:
                  code: "UNAUTHORIZED"
                  message: "Invalid or expired token"
...
    RateLimited:
      description: Too many requests
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying
        X-RateLimit-Limit:
          schema:
            type: integer
          description: Rate limit (requests per time window)
        X-RateLimit-Remaining:
          schema:
            type: integer
          description: Remaining requests in current window
        X-RateLimit-Reset:
          schema:
            type: integer
          description: Unix timestamp when rate limit resets
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            rate_limited:
              summary: Rate limit exceeded
              value:
                error:
                  code: "RATE_LIMITED"
                  message: "Too many requests. Please try again later."
  schemas:
...
    SuccessResponse:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
        data:
          type: object
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum: 
                - INVALID_INPUT
                - UNAUTHORIZED
                - FORBIDDEN
                - NOT_FOUND
                - DUPLICATE_ENTRY
                - LIMIT_EXCEEDED
                - INVALID_OPERATION
                - CONFLICT
                - PRECONDITION_FAILED
                - TIMEOUT
                - INTERNAL_ERROR
                - RATE_LIMITED
              description: |
                Error codes:
                - INVALID_INPUT: Validation failed or malformed request
                - UNAUTHORIZED: Missing or invalid authentication
                - FORBIDDEN: Insufficient permissions
                - NOT_FOUND: Resource not found or cross-tenant access
                - DUPLICATE_ENTRY: Unique constraint violation
                - LIMIT_EXCEEDED: Subscription or rate limit exceeded
                - INVALID_OPERATION: Operation not allowed in current state
                - CONFLICT: Resource conflict (e.g., concurrent modification)
                - PRECONDITION_FAILED: Precondition not met
                - TIMEOUT: Request timeout
                - INTERNAL_ERROR: Unexpected server error
                - RATE_LIMITED: Too many requests
            message:
              type: string
              description: Human-readable error message
              example: "Validation failed"
            details:
              type: array
              description: Detailed validation errors (for INVALID_INPUT)
              items:
                type: object
                properties:
                  field:
                    type: string
                    description: Field name that failed validation
                  message:
                    type: string
                    description: Specific validation error message
      example:
        error:
          code: "INVALID_INPUT"
          message: "Validation failed"
          details:
            - field: "email"
              message: "Must be a valid email address"
            - field: "password"
              message: "Must be at least 8 characters"
...
    RegisterRequest:
      type: object
      required: [organization_name, subdomain, full_name, email, password]
      properties:
        organization_name:
          type: string
          minLength: 2
          maxLength: 100
          example: "Acme Corp"
        subdomain:
          type: string
          pattern: ^[a-z0-9-]{3,32}$
          example: "acme"
        full_name:
          type: string
          minLength: 2
          maxLength: 100
          example: "John Doe"
        email:
          type: string
          format: email
          maxLength: 255
          example: "john@acme.com"
        password:
          type: string
          minLength: 8
          maxLength: 128
          format: password
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
          description: Valid refresh token issued by login/register/refresh
    AuthSuccessResponse:
      type: object
      required: [access_token, refresh_token, token_type, expires_in, user]
      properties:
        access_token:
          type: string
          description: |
            JWT access token for API authentication.
            
            **Token Structure:**
            - Header: {"alg": "HS256", "typ": "JWT"}
            - Payload: {"user_id": "...", "org_id": "...", "role": "...", "exp": ..., "iat": ...}
            - Signature: HMACSHA256(base64UrlEncode(header) + "." + base64UrlEncode(payload), secret)
            
            **Expiry:** 1 hour (3600 seconds)
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiNTUwZTg0MDAtZTI5Yi00MWQ0LWE3MTYtNDQ2NjU1NDQwMDAwIiwib3JnX2lkIjoiNjYwZTg0MDAtZTI5Yi00MWQ0LWE3MTYtNDQ2NjU1NDQwMDAwIiwicm9sZSI6Ik9XTkVSIiwiZXhwIjoxNzE1Nzg0MDAwLCJpYXQiOjE3MTU3ODA0MDB9.signature"
        refresh_token:
          type: string
          description: |
            Long-lived refresh token for obtaining new access tokens.
            
            **Expiry:** 7 days (604800 seconds)
            **Usage:** POST to /auth/refresh with this token to get new access token
            **Security:** Invalidated on password change or logout
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiNTUwZTg0MDAtZTI5Yi00MWQ0LWE3MTYtNDQ2NjU1NDQwMDAwIiwidHlwZSI6InJlZnJlc2giLCJleHAiOjE3MTYzODUyMDAsImlhdCI6MTcxNTc4MDQwMH0.signature"
        token_type:
          type: string
          enum: [Bearer]
          description: Token type for Authorization header
          example: "Bearer"
        expires_in:
          type: integer
          description: Access token expiry time in seconds
          example: 3600
        user:
          allOf:
            - $ref: '#/components/schemas/User'
            - type: object
              required: [role, organization_id]
              properties:
                role:
                  $ref: '#/components/schemas/Role'
                organization_id:
                  type: string
                  format: uuid
    TokenPair:
      type: object
      required: [access_token, refresh_token]
      properties:
        access_token:
          type: string
          description: Short-lived JWT (15 min default)
        refresh_token:
          type: string
          description: Long-lived refresh token (7 days default)
    TokenPairResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/TokenPair'
    Role:
      type: string
      enum: [OWNER, ADMIN, MEMBER]

    # ------ User / Organization ------
    User:
      type: object
      required: [id, email, full_name, role, organization_id, created_at]
      properties:
        id:
          type: string
          format: uuid
          description: Unique user identifier
          example: "550e8400-e29b-41d4-a716-446655440000"
        email:
          type: string
          format: email
          maxLength: 255
          description: User's email address (unique within organization)
          example: "john@acme.com"
        full_name:
          type: string
          minLength: 1
          maxLength: 100
          description: User's full name
          example: "John Doe"
        role:
          $ref: '#/components/schemas/Role'
        organization_id:
          type: string
          format: uuid
          description: Organization the user belongs to
          example: "660e8400-e29b-41d4-a716-446655440000"
        must_reset_password:
          type: boolean
          description: Whether user must reset password on next login
          default: false
        created_at:
          type: string
          format: date-time
          description: User creation timestamp
          example: "2026-05-01T10:00:00Z"
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
          example: "2026-05-15T14:30:00Z"
    Organization:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        subdomain:
          type: string
        subscription_plan:
          type: string
          enum: [FREE, LITE, PRO]
        created_at:
          type: string
          format: date-time
    UserWithOrgResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                user:
                  $ref: '#/components/schemas/User'
                organization:
                  $ref: '#/components/schemas/Organization'
                role:
                  $ref: '#/components/schemas/Role'

    # ------ Divisions ------
...